Bump to toml v0.9.5
This includes breaking API changes, but no changes to output formatting
that are bad enough to break my (basic) tests. It has a new parser and
writer, and came out in early July, so it seems like there's been enough
time to shake out the major bugs. However, whenever I get around to the
xt v0.20.0 release, it might be good to check on this one again.