Put pipecheck in the BufWriter, not the other way around
EPIPE can only happen on writes to the real stderr, so checking for
those failures on every _buffered_ write is silly.

This is also my justification for not cargo-culting `#[cold]` or other
silly "performance" attributes on pipecheck's termination functions.
This only matters when you're making a system call, at which point
you're doing something far more expensive than a branch misprediction
or whatever.